requirements.txt
fastapi
httpx
pydantic  
psycopg2-binary
python-dotenv
pytest-dotenv
pyparsing
pytest